rm_old_xml files() -> Improved the function with warnings and make sure
it deletes only the html files inside the tour folder

+++ newvt.sh
@@ -891,11 +891,18 @@ add_list() {
 }
 
 rm_old_xml_files() {
-
     if [ -d "$scenes_dir" ]; then
-        find $1 -type f \( -iname "*.html" ! -iname "template.html" ! -iname "index.html" \) -exec rm -rf {} \;
+        if [ ! -z $1 ]; then
+            find $1 -type f -name "*.html" -exec rm -rf {} \;
+            # Remove any xml with date stamp: year 2000 onwards
             find $1 -name "tour20*.xml" -exec rm -rf {} \;
             find $1 -name "tour_clean20*.xml" -exec rm -rf {} \;
+        else
+            echo_attention "rm_old_xml_files() -> \$1 not defined"
+            echo $1
+        fi
+    else
+        echo_warning "rm_old_xml_files() -> $scenes_dir folder doesn't exist"
     fi
 }
